users crud created

+++ database/migrations/2024_06_27_113019_drop_designation_column_from_job_listings_table.php
@@ -12,7 +12,10 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('job_listings', function (Blueprint $table) {
-            $table->dropColumn('designation');
+            // Check if column exists before dropping
+            if (Schema::hasColumn('job_listings', 'designation')) {
+                $table->dropColumn('designation');
+            }
         });
     }
